OfficeXlsFile2 linux build - complete
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@64139 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeXlsFile2/source/XlsFormat/Binary/CFRecord.h
@@ -114,10 +114,9 @@ public:
     CFRecord& operator>>(unsigned char& val)	{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(unsigned short& val)	{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(unsigned int& val)		{ loadAnyData(val);	return *this; }
-    CFRecord& operator>>(long& val)				{ loadAnyData(val);	return *this; }
+    CFRecord& operator>>(int& val)				{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(double& val)			{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(_GUID_& val)			{ loadAnyData(val);	return *this; }
-    //CFRecord& operator>>(unsigned int& val)	{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(short& val)			{ loadAnyData(val);	return *this; }
     CFRecord& operator>>(char& val)				{ loadAnyData(val);	return *this; }
 	CFRecord& operator>>(bool& val);
@@ -125,10 +124,9 @@ public:
     CFRecord& operator<<(unsigned char& val)	{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(unsigned short& val)	{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(unsigned int& val)		{ storeAnyData(val);	return *this; }
-    CFRecord& operator<<(long& val)				{ storeAnyData(val);	return *this; }
+    CFRecord& operator<<(int& val)				{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(double& val)			{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(_GUID_& val)			{ storeAnyData(val);	return *this; }
-    //CFRecord& operator<<(unsigned int& val)	{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(short& val)			{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(char& val)				{ storeAnyData(val);	return *this; }
     CFRecord& operator<<(wchar_t& val)			{ storeAnyData(val);	return *this; }
